Set organization domain in solarus-launcher to ensure that the QtWayland backend reports a valid AppId
Without this, it is not possible to have a non-generic window icon displayed under Wayland.

Nit: Since the domain name contains a dash, the reported AppId will *still* not be valid, but this will be worked around in specific distributions (Flatpak) until a proper fix has been submitted to QtWayland.